+++ web/src/widget/label.dart
@@ -18,17 +18,24 @@
 
 library summary;
 
-import 'dart:html' as dart_html;
-
 import 'package:dart_web_toolkit/ui.dart' as ui;
 
 void main() {
-  ui.SimplePanel panel = new ui.SimplePanel();
-  panel.setSize("200px", "30px");
-
-  ui.Hyperlink widget = new ui.Hyperlink("Home Page", false, "Home");
-  panel.addStyleName("demo-panel");
-  panel.add(widget);
   
-  ui.RootPanel.get().add(panel);
+  ui.Grid grid = new ui.Grid(3, 2);
+  
+  grid.setWidget(0, 0, new ui.Label("Label:"));
+  grid.setWidget(0, 1, new ui.Label("Casual text"));
+  
+  grid.setWidget(1, 0, new ui.Label("Date Label:"));
+  ui.DateLabel dLabel = new ui.DateLabel();
+  dLabel.setValue(new DateTime.now());
+  grid.setWidget(1, 1, dLabel);
+  
+  grid.setWidget(2, 0, new ui.Label("Number Label:"));
+  ui.NumberLabel nLabel = new ui.NumberLabel();
+  nLabel.setValue(123.12);
+  grid.setWidget(2, 1, nLabel);
+ 
+  ui.RootPanel.get().add(grid);
 }